Add a horizontal scrollbar component (#212)
Added a horizontal scrollbar components.

This component acts similar to a beam in a frame, except that it can have an arbitrary slider position set as some kind of fraction of the progress through the scrolled component.  It can be associated with a component to follow its focus and highlight appropriately.  Clicking on either side of the slider sends a scroll message.