Implement scrolling areas (#216)
Introduces scroll_pane, a class that wraps a tracked component into a viewport and implements vertical and horizontal scrollbars to track its positions.

Also changes edit controls so that cursors do not leave the bounds of the edit. The position of the caret is instead tracked separately. The control flow for a viewport observing such a control is thus that they receive a preferred size change and grant it, which causes the cursor to be updated to the caret position, which may cause the viewport to re-anchor.